deployments/requirements.txt
# AWS CDK dependencies for User Management Service infrastructure
aws-cdk-lib>=2.0.0
constructs>=10.0.0
